PXB-1947: backup crash when keyring settings are incorrect
Problem:

When xtrabackup is making a backup and it cannot decrypt tablespace
header it is looking if tablespace header encryption info was already
parsed by the log copying thread, in this case xtrabackup will use the
key extracted from the redo log since tablespace keys never changed.

The issue is that when keyring settings are incorrect, and log copying
thread did not yet extracted any key information,
recv_sys->encryption_list is NULL.

Fix:

Check if encryption_list is not NULL before trying to iterate it.